+++ src/library/ShimmedStabilityAIClient/definition.ts
